added userId to WSMessages
