+++ src/BoardStateMachine/userInputHandlers/handleKeyboard.ts
@@ -1,5 +1,6 @@
 import { KeyboardEvent } from 'react';
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 function handleKeyboard(e: KeyboardEvent): void {
     // console.log(e);
 }
