+++ src/components/common/StylesMenu/MenuOptions/ColorSelector/ColorSelector.tsx
@@ -9,7 +9,7 @@ interface ColorSelector {
     onChange(value: string, key: string): void;
 }
 
-const ColorSelector = ({ onChange, styleKey, color }: ColorSelector): React.ReactElement => {
+const ColorSelector = ({ onChange, styleKey }: ColorSelector): React.ReactElement => {
     const handleClick = (e: React.MouseEvent<HTMLButtonElement>) => {
         const { value } = e.currentTarget;
         onChange(value, styleKey);
